[hal] Further split pinmap into each platform
