+++ hal/src/gcc/pinmap_impl.h
@@ -1,6 +1,23 @@
+/*
+ * Copyright (c) 2018 Particle Industries, Inc.  All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation, either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, see <http://www.gnu.org/licenses/>.
+ */
+
 #pragma once
 
-#include "pinmap_hal.h"
+#if PLATFORM_ID == PLATFORM_GCC
 
 const pin_t TOTAL_PINS = 21;
 const pin_t TOTAL_ANALOG_PINS = 8;
@@ -53,14 +70,16 @@ const pin_t TIMER4_CH1 = 1;
 const pin_t TIMER4_CH2 = 0;
 
 // SPI pins
-const pin_t SS   = 12;
-const pin_t SCK  = 13;
+
+const pin_t SS = 12;
+const pin_t SCK = 13;
 const pin_t MISO = 14;
 const pin_t MOSI = 15;
 
 // I2C pins
-const pin_t SDA  = 0;
-const pin_t SCL  = 1;
+
+const pin_t SDA = 0;
+const pin_t SCL = 1;
 
 // DAC pins on Photon
 const pin_t DAC1 = 16;
@@ -68,3 +87,5 @@ const pin_t DAC2 = 13;
 
 const uint8_t LSBFIRST = 0;
 const uint8_t MSBFIRST = 1;
+
+#endif // PLATFORM_ID == PLATFORM_GCC
